extract droping function. fix autotest
