+++ connectivity/constants.py
@@ -1,3 +1,4 @@
 SDS011_MODEL = 2
 MOBILE_GPS = 3
 PING_MODEL = 1
+STATION_VERSION = "v0.8.0"
