+++ connectivity/src/stations/httpstation.py
@@ -4,15 +4,13 @@ import threading
 import time
 import json
 import cgi
-import copy
-import hashlib
 import typing as tp
 import logging.config
 from http.server import BaseHTTPRequestHandler, HTTPServer
 from prometheus_client import Gauge
 
-from ...constants import SDS011_MODEL, MOBILE_GPS
-from .istation import IStation, STATION_VERSION
+from .istation import IStation
+from ...constants import STATION_VERSION
 from ..sensors import EnvironmentalBox
 
 from connectivity.config.logging import LOGGING_CONFIG
@@ -94,22 +92,11 @@ class HTTPStation(IStation):
 
     def get_data(self) -> tp.List[dict]:
         global sessions
+        global thlock
         result = []
-        for k, v in self._drop_dead_sensors().items():
+        with thlock:
+            stripped = self.drop_dead_sensors(sessions)
+            ALIVE_SENSORS_METRIC.set(len(stripped))
+        for k, v in stripped.items():
             result.append(v)
         return result
-
-    def _drop_dead_sensors(self) -> dict:
-        global sessions
-        global thlock
-        stripped = dict()
-        current_time = int(time.time())
-        with thlock:
-            sessions_copy = copy.deepcopy(sessions)
-            for k, v in sessions_copy.items():
-                if (current_time - v.timestamp) < self.DEAD_SENSOR_TIME:
-                    stripped[k] = v
-                else:
-                    del sessions[k]
-        ALIVE_SENSORS_METRIC.set(len(stripped))
-        return stripped
